FIX: tests con imports erroneos
